Fixed open and close issue
